Added sending of LightSensor Data over Dash7
Added code to send the light sensor data over UART to the Dash7 node using a custom protocol. The definition of the protocol for the UART message is as follows:

Message Structure:
Each message will consist of the following fields:

Start Delimiter (1 byte): A unique byte to indicate the start of a message. Defined as 0xAA.
Sensor Type (1 byte): A byte to identify the type of sensor. Different sensor types will have different assigned values. For example, 0x01 for temperature, 0x02 for humidity, 0x03 for light, etc.
Timestamp (4 bytes): A 4-byte field representing the time of measurement. This could be the Unix timestamp (number of seconds since January 1, 1970).
Sensor Data (2 bytes): The actual sensor reading. It's a 2-byte field representing the sensor data.
Checksum (1 byte): A simple checksum for error checking. It could be the sum of all previous bytes in the message.
End Delimiter (1 byte): A unique byte to indicate the end of a message. Defined as 0x55.